Continued working on Bank class and improved Chip class.
I have abandoned the idea of 3 chip stacks and will stay with 1 chip max.

I still need to figure out whether I will be making and deleting chips, or just hiding them. I will probably be just hiding them, but then how will this work for the bet stack? Would I have 10 chips all invisible on top of each other and just make them appear depending on the chip clicked. Sounds like a good idea, but then I need to add a variable to keep track of the visible chip in order to easily disalbe it. It can be and int from 0 to 9 which would describe the place of that chip in the array. So I will have 2 arrays of 10 chips. One for the bank and other one for the bet. Chips will exist at all time and chips will hide when needed.

Next time I need to realize the above idea in the Bank class.

After that I will make a simple animation of a chip moving to and from the stack.

5/25/2022