I really dont like the fact that I cant use a method of a class inside that same class.
I will replace fillBank with makeBank method.

Fillbank method will instead be changed so it is only making the number of chips visible corresponding to the self.cash value and will hide the over cash chips.

Fillbank will be called every time a chip is clicked, whereas the makeBank is only called once when the bank is created.

I might want to make a function in BlackJack.py that creates and returns the bank.

showBank and hideBank functions will be resposible for the sceen change (bet to deal and vice versa)

I need to shift the bet stack slightly to the left and show the bet value beside it (at all times). Cash value will only be visible during bet sceene.

Add a function for the future withdrawl button (finishes the game and takes bet + cash)

add variable to keep track of all time high (compare to the cash value at the very start of every bet sceene - that is when self.bet = 0 - )

Turn removeChip and addChip into animation functions

NEXT TIME I SHOULD FINISH AND TEST THE READ CHIP FUNCTION (rename it to readBank)

readChip will be reading the mouse, manipulating chips if they are clicked or hideBank and showDeal if the deal button is clicked (make deal button) .

5/26/2022