Commented and improve Hand class. Made Abstract Class.
Next time add comments to the Dealer class and keep working on the splitting algorithms.

Probably make more methods for when the Stand button is clicked. Check if there are unplayed hands in the self.splitHands. If no unplayed hands, open dealer card. Do dealer draw and cycle through every hand in self.splitHands comparing that hand to the dealers and showing the result to the user (win, loose, push). Modify the bank.cash correspondigly.

Oh, and before splitting check if the user has enough money. Then show the split button.

6/9/2022